# backend/app/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: Online Examination Platform API
  version: 1.0.0
  description: Documentation for the existing Online Examination Platform API.
servers:
  - url: http://127.0.0.1:5000
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/exams/{exam_id}/questions:
    get:
      tags: [Student]
      summary: Retrieve questions for an examination
      x-required-role: STUDENT
      parameters: [{ $ref: '#/components/parameters/ExamId' }]
      responses: { '200': { $ref: '#/components/responses/Success' }, '401': { $ref: '#/components/responses/Error' }, '403': { $ref: '#/components/responses/Error' }, '404': { $ref: '#/components/responses/Error' } }
    post:
      tags: [Faculty]
      summary: Add a question to an exam
      x-required-role: FACULTY
      parameters: [{ $ref: '#/components/parameters/ExamId' }]
      requestBody:
        required: true
        content:
          application/json: { schema: { type: object, additionalProperties: true } }
      responses: { '201': { $ref: '#/components/responses/Success' }, '400': { $ref: '#/components/responses/Error' }, '401': { $ref: '#/components/responses/Error' }, '403': { $ref: '#/components/responses/Error' } }
  /api/v1/attempts/{attempt_id}/next-question:
    get:
      tags: [Student]
      summary: Retrieve the next adaptive question
      x-required-role: STUDENT
      parameters: [{ $ref: '#/components/parameters/AttemptId' }]
      responses: { '200': { $ref: '#/components/responses/Success' }, '401': { $ref: '#/components/responses/Error' }, '500': { $ref: '#/components/responses/Error' } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ExamId: { name: exam_id, in: path, required: true, schema: { type: integer, format: int64 } }
    AttemptId: { name: attempt_id, in: path, required: true, schema: { type: integer, format: int64 } }
...
  schemas:
...
    ApiResponse:
      type: object
      properties:
        success: { type: boolean }
        message: { type: string }
        data: { nullable: true }
    ErrorResponse:
      type: object
      properties:
        success: { type: boolean, example: false }
        message: { type: string }
  responses:
    Success:
      description: Successful API response
      content:
        application/json: { schema: { $ref: '#/components/schemas/ApiResponse' } }
...
    Error:
      description: Error response
      content:
        application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } }